tui: reduce legacy core helper usage
