+++ codex-rs/tui/src/app/startup_prompts.rs
@@ -66,9 +66,9 @@ pub(super) fn emit_project_config_warnings(app_event_tx: &AppEventSender, config
 }
 
 pub(super) fn emit_system_bwrap_warning(app_event_tx: &AppEventSender, config: &Config) {
-    let Some(message) = crate::legacy_core::config::system_bwrap_warning(
-        config.permissions.permission_profile.get(),
-    ) else {
+    let Some(message) =
+        codex_sandboxing::system_bwrap_warning(config.permissions.permission_profile.get())
+    else {
         return;
     };
