+++ codex-rs/utils/cli/src/lib.rs
@@ -1,6 +1,7 @@
 mod approval_mode_cli_arg;
 mod config_override;
 pub(crate) mod format_env_display;
+mod resume_command;
 mod sandbox_mode_cli_arg;
 mod shared_options;
 
@@ -8,5 +9,6 @@ pub use approval_mode_cli_arg::ApprovalModeCliArg;
 pub use codex_protocol::config_types::ProfileV2Name;
 pub use config_override::CliConfigOverrides;
 pub use format_env_display::format_env_display;
+pub use resume_command::resume_command;
 pub use sandbox_mode_cli_arg::SandboxModeCliArg;
 pub use shared_options::SharedCliOptions;
